feat: Update Dependencies
**Updated Dependencies:**

- **numpy**: 1.26.4 → 2.1.2
- scipy: 1.13.1 → 1.14.1
- pre-commit: 3.7.1 → 4.0.1
- ruff: 0.4.7 → 0.6.9
- pytest: 8.2.2 → 8.3.3
- coverage: 7.5.3 → 7.6.2
- sphinx: 7.3.7 → 8.1.0
- numpydoc: 1.7.0 → 1.8.0
- myst-nb: 1.1.0 → 1.1.2
- sphinxcontrib-bibtex: 2.6.2 → 2.6.3
- matplotlib: 3.8.4 → 3.9.2

Performance improvements in **numpy 2.0.0**:

- Sorting functions (sort, argsort, partition, argpartition) have been accelerated through the use of the Intel x86-simd-sort and Google Highway libraries, and may see large (hardware-specific) speedups.

+++ infomeasure/measures/utils/symbolic.py
@@ -66,7 +66,7 @@ def permutation_to_integer(perm: ndarray, dtype: type = uint64) -> int:
     for i in range(n):
         factoradic[i] = np_sum(perm[i] > perm[i + 1 :], dtype=dtype)
     integer = np_sum(
-        factoradic * array([factorial(n - 1 - i) for i in range(n)], dtype=dtype),
+        factoradic * array([factorial(n - 1 - i) for i in range(n)]),
         dtype=dtype,
     )
     return integer
